Tests: compute refs in FP64 (#6)
* fix two tests that had same name

* ref in fp64

+++ cuequivariance_torch/tests/operations/tp_fully_connected_test.py
@@ -52,6 +52,11 @@ def test_fully_connected(
     if layout == cue.mul_ir:
         d = d.add_or_transpose_modes("uvw,ui,vj,wk+ijk")
     mfx = cuet.TensorProduct(d).cuda()
-    out2 = mfx(m.weight, x1, x2, use_fallback=True)
+    out2 = mfx(
+        m.weight.to(torch.float64),
+        x1.to(torch.float64),
+        x2.to(torch.float64),
+        use_fallback=True,
+    ).to(out1.dtype)
 
     torch.testing.assert_close(out1, out2, atol=1e-5, rtol=1e-5)
